replace cli.choose by cli.run and fname by level
This patch replaces cli.choose by cli.run to make the script compatible with
the automation framework for examples.nutils.org. To make this work, the firt
argument of the main function has been repurposed to select either a single
level or a range. The change from fname to level for the main function is a
step down in generality, but the requirements of inner and outer boundary
groups anyhow meant that not just any msh file could be inserted, so it would
not appear to detract much from the purpose of the script.